deploy: improve mem handling for video tests
